+++ GithubUsers/Domain/Services/Users/FetchUsersService.swift
@@ -17,7 +17,7 @@ actor FetchUsersService {
 
 // MARK: - UsersFetchable
 extension FetchUsersService: UsersFetchable {
-    func fetchUsers(perPage: Int, since: Int) async throws -> [User] {
+    func fetchUsers(perPage: Int, since: Int64) async throws -> [User] {
         let requestData = UsersEndPoint.getUsersWith(perPage: perPage, since: since)
 
         do {
